Factor base element fixup into HTMLUtil
This needs to be fixed both in macros that show HTML frames
and in the stat sheets, so there should be a common util for it.
