+++ src/main/java/net/rptools/maptool/client/functions/MacroDialogFunctions.java
@@ -35,13 +35,11 @@ import net.rptools.maptool.language.I18N;
 import net.rptools.maptool.model.library.Library;
 import net.rptools.maptool.model.library.LibraryManager;
 import net.rptools.maptool.util.FunctionUtil;
+import net.rptools.maptool.util.HTMLUtil;
 import net.rptools.parser.Parser;
 import net.rptools.parser.ParserException;
 import net.rptools.parser.VariableResolver;
 import net.rptools.parser.function.AbstractFunction;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Element;
-import org.jsoup.parser.Tag;
 
 public class MacroDialogFunctions extends AbstractFunction {
   private static final MacroDialogFunctions instance = new MacroDialogFunctions();
@@ -200,23 +198,7 @@ public class MacroDialogFunctions extends AbstractFunction {
             I18N.getText("macro.function.html5.invalidURI", url.toExternalForm()));
       }
 
-      htmlString = library.get().readAsString(url).get();
-
-      var document = Jsoup.parse(htmlString);
-      var head = document.select("head").first();
-      if (head != null) {
-        String baseURL = url.toExternalForm().replaceFirst("\\?.*", "");
-        baseURL = baseURL.substring(0, baseURL.lastIndexOf("/") + 1);
-        var baseElement = new Element(Tag.valueOf("base"), "").attr("href", baseURL);
-        if (head.children().isEmpty()) {
-          head.appendChild(baseElement);
-        } else {
-          head.child(0).before(baseElement);
-        }
-
-        htmlString = document.html();
-      }
-
+      htmlString = HTMLUtil.fixHTMLBase(library.get().readAsString(url).get(), url);
     } catch (InterruptedException | ExecutionException | IOException e) {
       throw new ParserException(e);
     }
